Selection Field Fix (#2828)
* bugbash fixes

* bug bash fixes

* version upgrade

* version upgrade for ui-components

* HCMPRE-2933

* Added RenderSelectionField

---------
